spaceorbiterclickplanets/spaceorbiter1clickplanets/backend/requirements.txt
axios==0.4.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
Pygments==2.17.2
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.31.0
rich==13.7.0
urllib3==2.2.0
Werkzeug==3.0.1
